requirements.txt
Django==5.2.5
groq==0.11.0
python-docx==1.1.2
markdown
gunicorn
whitenoise[brotli]
psycopg2-binary
python-decouple
dj-database-url
